Refactor integration tests using Arquillian's @InSequence.
Some integration tests are verifying that resources are cleaned up on session expiration.
Instead of calling tests from other tests (which is slow and less readable), we now invoke
tests in a deterministic order so that a test for cleaning up can run immediately after a test for setting up.